Se modifico el microservicio de Usuarios (Merge rama gabriel)

+++ Usuarios/src/main/java/com/example/USER/repository/UsuarioRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository <Usuarios,Integer> {
+public interface UsuarioRepository extends JpaRepository <Usuarios, Long> {
 
 }
